POC of inline diagnostics below errors
